Removed redundant logics regarding building executable paths, adding better error catching logic, norming needs to be done again
