+++ include/execution.h
@@ -6,7 +6,7 @@
 /*   By: jteissie <jteissie@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/07 16:53:36 by jteissie          #+#    #+#             */
-/*   Updated: 2024/07/12 14:19:15 by jteissie         ###   ########.fr       */
+/*   Updated: 2024/07/12 15:35:16 by jteissie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,22 +22,16 @@
 # define PATH_ERROR 127
 # define EXIT_FAILURE 1
 
-char	*get_execpath(char **envp, char *av);
-void	cat_slash(char **envp);
 void	trash(char **array);
-void	copycat_str(char *out, char *cpy_src, char *cat_src);
-char	*get_path(char *av, char **env);
-char	*ft_str_realloc_join(char *stash, char *add);
 void	handle_error(char *message, int code);
-void	try_absolute_path(char **command, char **env);
-void	execute(char *av, char **env);
+void	execute_cmd(char *cmd, char **env);
 
 int		open_files(int file_fd[], t_lex_parser *table);
 void	get_redirections(t_lex_parser *table, char *redirection[]);
 void	go_to_first_table(t_lex_parser *roaming, t_lex_parser *parsed);
 
 int		execute_commands(t_lex_parser *tables, char **envp);
-void	process_command(t_lex_parser *parsed, char **envp);
-void	redirect_child(int file_fd[], int p_fd[]);
+int		process_command(t_lex_parser *parsed, char **envp);
+int		redirect_child(int file_fd[], int p_fd[]);
 
 #endif
